Move dict_conversion utility out of io package
This is to get #205 to work, since doing this removes the circular
imports. But the dict_conversion tool is more connected to #119

Also, dict_conversion never belonged in io because io is about
*hardware* io and not *filesystem* io.
