+++ src/openlifu/bf/sequence.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 import pandas as pd
 
-from openlifu.io.dict_conversion import DictMixin
+from openlifu.util.dict_conversion import DictMixin
 
 
 @dataclass
